Bugfixes - Changed names of references in new code to match the existing code
Changed the names of references in the new code to match the names of the objects and fields being referenced in our existing code. I don't fel bad doing this, because Jame's code did not refer to these objects in a consistent way across different classes in his own code. (He calls a sensor "is" in the ExitController method, and a similar sensor "outsideEntrySensor_" in the EntryController method.