Fixed a bug where the exit controller would leave the blocked state when certain actions were performed.
Fixed a bug where the exit controller would leave the blocked state after the Ticket Inserted method and ticket-taken methods were called while in the blocked state, returning to idle while the outside senosr was still blocked. The fix was simply to manually chack the outside sensor after a rejected ticket was taken.