Fixed a bug found in User Acceptance Testing
Fixed a bug where a user could back out of entering the carpark but would still be able to pay their ticket. The root cause was the ticket having .enter() called both in the factory and when actually entering the carpark. The method call was removed from the factory, and a check was added to the PaystationController to make sure that the ticket did not have the default value for getEntryDateTime.